% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-path.R
\name{add_path}
\alias{add_path}
\title{Concat the categories of a variable}
\usage{
add_path(res_gda_quali, var, var_levels, exclude = NULL, axes = 1:2,
  linetype = "dashed", colour = "black", size = 1)
}
\arguments{
\item{res_gda_quali}{GDA result.}

\item{var}{variable name.}

\item{var_levels}{categories to concat.}

\item{exclude}{categories to exclude.}

\item{axes}{axes to plot.}

\item{linetype}{linetype of concat path.}

\item{colour}{colour of concat path.}

\item{size}{size of concat path.}
}
\value{
ggplo2 path geom.
}
\description{
Concat the categories of a variable
}

