% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-load.R
\name{db_load}
\alias{db_load}
\title{Save data frame as SQLite table.}
\usage{
db_load(remote_db_path)
}
\arguments{
\item{remote_db_path}{}
}
\value{
The SQLite database.
}
\description{
Save data frame as SQLite table.
}
\examples{
db_load('/path/to/database.sqlite')
}

