% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim-desc.R
\name{dim_desc}
\alias{dim_desc}
\title{Dimension description of component analysis.}
\usage{
dim_desc(res, axes = 1:3, proba = 0.05)
}
\arguments{
\item{proba}{}
}
\value{
list of representative kategories and modalities.
}
\description{
Dimension description of component analysis.
}

