% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimdesc.R
\name{dimdesc}
\alias{dimdesc}
\title{Dimension description of component analysis.}
\usage{
dimdesc(res, axes = 1:3, proba = 0.05)
}
\arguments{
\item{proba}{}
}
\value{
list of representative kategories and modalities.
}
\description{
Dimension description of component analysis.
}

