% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excl-mfa-group.R
\name{excl_mfa_group}
\alias{excl_mfa_group}
\title{Extract list of MFA group indices to exclude (for sMFA).}
\usage{
excl_mfa_group(df_mfa, group_mfa, pattern)
}
\arguments{
\item{df_mfa}{MFA optmised data frame.}

\item{group_mfa}{MFA group definition.}

\item{pattern}{search pattern (regular expression).}
}
\value{
list with MFA group specific exclude indices.
}
\description{
Extract list of MFA group indices to exclude (for sMFA).
}

