% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-legend.R
\name{extract_legend}
\alias{extract_legend}
\title{Extract legends form ggplot2 objects.}
\usage{
extract_legend(p)
}
\arguments{
\item{p}{ggplot2 object (plot) containing legends (guides).}
}
\value{
plottable legend grob.
}
\description{
Extract legends form ggplot2 objects.
}
