% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz_add_sup_ind.R
\name{fviz_add_sup_ind}
\alias{fviz_add_sup_ind}
\title{Add supplementary individuals.}
\usage{
fviz_add_sup_ind(res_gda, excl = NULL, sup_ind = NULL, colour = "red",
  label = NULL, size = 10, myriad = TRUE)
}
\arguments{
\item{res_gda}{GDA results.}

\item{excl}{Exclude modalities (list).}

\item{sup_ind}{supplementary individual profiles (data.frame).}

\item{colour}{colour of point and labels.}

\item{label}{label names.}

\item{size}{label size.}
}
\value{
ggplot2 visalization of supplementary individuals.
}
\description{
Add supplementary individuals.
}

