% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-barplot.R
\name{fviz_barplot}
\alias{fviz_barplot}
\title{Visualize a barplot.}
\usage{
fviz_barplot(dfname, xlab = "", ylab = "", sort = FALSE,
  relative = FALSE, labels = FALSE, labels_inline = TRUE,
  amount = FALSE, rotate = FALSE, title = "", textsize = 14,
  titlesize = 1.5 * textsize, labelsize = textsize/3, drop = FALSE,
  digits = 0, addsymb = "", ylim = NA, xlim = NA)
}
\arguments{
\item{dfname}{categorical variable.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{sort}{sort bars (boolean).}

\item{relative}{realtive values (boolean).}

\item{labels}{label bars (boolean).}

\item{labels_inline}{inline labels (boolean).}

\item{amount}{include total amount of observations (boolean).}

\item{rotate}{rotate bars (boolean).}

\item{title}{plot title.}

\item{textsize}{sife of axes texts.}

\item{titlesize}{title text size.}

\item{labelsize}{label text size.}

\item{drop}{drop missing data (boolean).}

\item{digits}{amount of label value digits.}

\item{addsymb}{suffix symbol.}

\item{ylim}{y-axis range.}

\item{xlim}{y-axis range.}
}
\value{
ggplot2 barplot.
}
\description{
Visualize a barplot.
}

