% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-dendrogram.R
\name{fviz_dendrogram}
\alias{fviz_dendrogram}
\title{Visualize HCPC hclus trees.}
\usage{
fviz_dendrogram(res_hcpc, palette = NULL, cluster = 1, labels = FALSE,
  circle = FALSE, hline = 0.8, pointsize = 2, linetype = "dashed",
  myriad = TRUE, cut_height = NULL)
}
\arguments{
\item{res_hcpc}{(s)HCPC results.}

\item{palette}{colour definition per cluster.}

\item{cluster}{amount of clusters.}

\item{labels}{plot labels (boolean).}

\item{circle}{plot circle (boolean).}

\item{hline}{hline height.}

\item{pointsize}{leaves pointsize.}

\item{linetype}{hline linetype.}

\item{myriad}{use Myriad Pro font (boolean).}

\item{cut_height}{cut dendrogram at specific hight.}
}
\value{
ggplot2 dendrogram visualization.
}
\description{
Visualize HCPC hclus trees.
}
