% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-hmfa-trajectory.R
\name{fviz_gda_hmfa_trajectory}
\alias{fviz_gda_hmfa_trajectory}
\title{Visualization of trajectories (connected HMFA partial points).}
\usage{
fviz_gda_hmfa_trajectory(res_gda, clust, select = list(name = NULL, contrib =
  NULL, cos2 = NULL, within_inertia = NULL), ellipse_level = 0.8647,
  ellipse_alpha = 0.1, axes = 1:2, myriad = TRUE, ellipses = FALSE,
  facet = FALSE, mean_path = FALSE)
}
\arguments{
\item{res_gda}{HMFA result (rownames have to be questionnaire IDs).}

\item{clust}{HCPC result of primary MFA.}

\item{select}{contribution, cos2, vector of names or within_inertia of individuals selection (within_inertia: vector containing the number of high variation and low variationindividuals).}

\item{ellipse_level}{the ellipse level. Default: 86.47\%.}

\item{ellipse_alpha}{opacity value.}

\item{axes}{axes to plot.}

\item{myriad}{use Myriad Pro font (boolean).}

\item{ellipses}{plot concentration ellipses (boolean).}

\item{facet}{plot ellipses/ individuals per year (boolean).}

\item{mean_path}{plot mean path (boolean). If yes, no ellipses and no facets are plotted.}
}
\value{
HMFA trajectory ggplot2 visualization.
}
\description{
Visualization of trajectories (connected HMFA partial points).
}

