% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-quali-ellipses.R
\name{fviz_gda_quali_ellipses}
\alias{fviz_gda_quali_ellipses}
\title{Title}
\usage{
fviz_gda_quali_ellipses(res_gda, df_var_quali, var_quali,
  title = "MCA quali var ellipses", facet = TRUE, alpha_point = 0.75,
  conc_linetype = "solid", conf_linetype = "solid",
  scale_mean_points = TRUE, axes = 1:2, palette = "Set1", myriad = TRUE,
  impute = TRUE, concentration_ellipses = TRUE,
  confidence_ellipses = FALSE, conf_colour = FALSE,
  plot_modif_rates = TRUE, ncol = 3, individuals = TRUE, impute_ncp = 2,
  relevel = NULL, alpha_ellipses = 0.15, plot_eta2 = TRUE,
  axis_lab_name = "Achse", show_prop = FALSE)
}
\arguments{
\item{res_gda}{GDA (MCA, MFA) result (rownames have to be individual questionnaire IDs).}

\item{df_var_quali}{data frame of one quali variable.}

\item{var_quali}{name if quali variable.}

\item{title}{plot title.}

\item{facet}{whether facet ellipses or not (boolean).}

\item{alpha_point}{opacity of individual points.}

\item{conc_linetype}{linetype of concentration ellipses.}

\item{conf_linetype}{linetype of confidence ellipses.}

\item{scale_mean_points}{scale mean point size in respect of the group size (boolean).}

\item{axes}{the GDA dimensions to plot.}

\item{palette}{Colour brewer scale.}

\item{myriad}{use Myriad Pro font family (boolean).}

\item{impute}{impute missing data (boolean).}

\item{concentration_ellipses}{plot concentration ellipse (boolean).}

\item{confidence_ellipses}{plot confidence ellipses (boolean).}

\item{conf_colour}{colour confidence ellipses (boolean).}

\item{plot_modif_rates}{plot modified rates instead of eigenvalue percentage (boolean).}

\item{ncol}{Number of facet columns.}

\item{individuals}{show individual points (boolean).}

\item{impute_ncp}{number of dimensions to predict missing values.}

\item{relevel}{character vector containing new level order.}

\item{alpha_ellipses}{concentration ellipses fill alpha.}

\item{plot_eta2}{plot eta2 value per axis (boolean).}

\item{axis_lab_name}{name of axis label.}

\item{show_prop}{show prop label as mean point (boolean).}
}
\value{
ggplo2 visualization with concentration and quali var ellipses.
}
\description{
Title
}
