% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-quali-supvar.R
\name{fviz_gda_quali_supvar}
\alias{fviz_gda_quali_supvar}
\title{Visualize supplementary variables.}
\usage{
fviz_gda_quali_supvar(res_gda, df_var_quali, var_quali_name,
  title = "MCA quali var structure", path = FALSE, linetype = "solid",
  scale_point = TRUE, axes = 1:2, scale_text = FALSE, palette = "Set1",
  na_exclude = TRUE)
}
\arguments{
\item{res_gda}{GDA result.}

\item{df_var_quali}{crossed variable data frame.}

\item{var_quali_name}{crossed variable name.}

\item{title}{plot title.}

\item{path}{plot path (boolean).}

\item{linetype}{specify path linetype.}

\item{scale_point}{scale points by weight (boolean).}

\item{axes}{which axes should be plotted.}

\item{scale_text}{scale text by weight (boolean).}

\item{palette}{RColorBrewer palette.}

\item{na_exclude}{handle missing data (boolean).}
}
\value{
ggplot2 visualization of supplementary variables.
}
\description{
Visualize supplementary variables.
}

