% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-trajectory.R
\name{fviz_gda_trajectory}
\alias{fviz_gda_trajectory}
\title{Visualization of trajectories (connected MFA active/passive individual points).}
\usage{
fviz_gda_trajectory(res_gda, select = list(name = NULL, within_inertia = NULL,
  case = NULL), axes = 1:2, myriad = TRUE, labels = FALSE)
}
\arguments{
\item{res_gda}{MFA or MCA result (rownames have to be questionnaire IDs including time number, e.g. 87654_1).}

\item{select}{vector of names, within_inertia of individuals selection (within_inertia: vector containing the number of high variation and low variationindividuals) or time_point (vector containing the semesters to plot).}

\item{axes}{axes to plot.}

\item{myriad}{use Myriad Pro font (boolean).}

\item{labels}{plot individual labels (boolean).}

\item{facet_labels}{rename facet labels (vector).}
}
\value{
HMFA trajectory ggplot2 visualization.
}
\description{
Visualization of trajectories (connected MFA active/passive individual points).
}

