% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-trajectory-ellipses.R
\name{fviz_gda_trajectory_ellipses}
\alias{fviz_gda_trajectory_ellipses}
\title{Title}
\usage{
fviz_gda_trajectory_ellipses(res_gda, df_var_quali, var_quali, axes = 1:2,
  myriad = TRUE, time_point_names = NULL)
}
\description{
Title
}

