% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-trajectory-quali.R
\name{fviz_gda_trajectory_quali}
\alias{fviz_gda_trajectory_quali}
\title{Title}
\usage{
fviz_gda_trajectory_quali(res_gda, df_var_quali, var_quali, axes = 1:2,
  myriad = TRUE, labels = FALSE, time_point_names = NULL,
  select = list(name = NULL, within_inertia = NULL, case = NULL))
}
\arguments{
\item{select}{}
}
\description{
Title
}

