% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-var.R
\name{fviz_gda_var}
\alias{fviz_gda_var}
\title{Visualize specific contributing modalities in a plane.}
\usage{
fviz_gda_var(res_gda, contrib = "auto",
  title = "GDA plane high contribution modalities", axes = 1:2,
  myriad = TRUE, group = NULL, group_names = NULL, group_style = "both",
  textsize = 4, colour_palette = "Dark2", individuals = FALSE,
  individuals_size = "auto", individuals_alpha = 0.5,
  individuals_names = FALSE, plot_modif_rates = TRUE)
}
\arguments{
\item{res_gda}{GDA result data frame.}

\item{contrib}{"auto" calculates the optimal modalities to show (based on the basic criterion). Otherwise define an amount of modalities to plot.}

\item{title}{plot title.}

\item{axes}{the GDA dimensions to plot.}

\item{myriad}{use Myriad Pro font (boolean).}

\item{group}{vector containing group definition.}

\item{group_names}{names of the groups.}

\item{group_style}{style to plot (vector containing "shape", "colour" or "both).}

\item{textsize}{size of the text.}

\item{colour_palette}{name of the used colour palette.}

\item{individuals}{show individual points/ biplot (boolean).}

\item{individuals_size}{set individual point size manual or "auto".}

\item{individuals_alpha}{set alpha value.}

\item{individuals_names}{plot individual names (boolean).}

\item{plot_modif_rates}{plot modified rates instead of eigenvalue percentage (boolean).}
}
\value{
ggplot2 visualization containing selected modalities.
}
\description{
Visualize specific contributing modalities in a plane.
}

