% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-var-axis.R
\name{fviz_gda_var_axis}
\alias{fviz_gda_var_axis}
\title{Visualize specific contributing modalities.}
\usage{
fviz_gda_var_axis(res_gda, axis = 1, contrib = 20,
  title = "GDA axis contribution 20")
}
\arguments{
\item{res_gda}{GDA result data frame.}

\item{axis}{dimension to be filtered.}

\item{contrib}{count of modalities to plot.}

\item{title}{plot title.}
}
\value{
ggplot2 visualization containing selected modalities.
}
\description{
Visualize specific contributing modalities.
}

