% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-mca-var-corr.R
\name{fviz_mca_var_corr}
\alias{fviz_mca_var_corr}
\title{Visualize MCA variable representation square.}
\usage{
fviz_mca_var_corr(res_gda, axes = c(1, 2), geom = c("point", "text"),
  labelsize = 4, pointsize = 2, invisible = NULL, labels = TRUE,
  repel = TRUE, select = list(name = NULL, eta2 = NULL), myriad = TRUE,
  plot_modif_rates = TRUE, title = "MCA - Variable Representation")
}
\arguments{
\item{res_gda}{MCA result.}

\item{axes}{axes to plot.}

\item{geom}{whether points or labels to plot.}

\item{labelsize}{size of labels.}

\item{pointsize}{size of points.}

\item{invisible}{hide "passive" or "active" variables.}

\item{labels}{label points or not (boolean).}

\item{repel}{repel labels (boolean).}

\item{select}{selection of variables (names) or eta2 values (all above value).}

\item{plot_modif_rates}{plot modified rates instead of eigenvalue percentage (boolean).}

\item{title}{plot title.}
}
\value{
ggplot2 visualization of variable correlation square (variables representation).
}
\description{
Visualize MCA variable representation square.
}

