% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-places-chronology.R
\name{fviz_places_chronology}
\alias{fviz_places_chronology}
\title{Title}
\usage{
fviz_places_chronology(data, id = "all", map, weekday = "all",
  timepattern = TRUE, distinct_places = FALSE,
  print_prop_duration = FALSE, legends = FALSE, structure = FALSE,
  title = "Orte Chronologie")
}
\arguments{
\item{structure}{}
}
\description{
Title
}

