% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gda-describe-group.R
\name{gda_describe_group}
\alias{gda_describe_group}
\title{Calculate group contributions.}
\usage{
gda_describe_group(res_gda, group = NULL, group_names = NULL, excl = NULL)
}
\arguments{
\item{res_gda}{MCA result.}

\item{group}{vector containing group definition.}

\item{group_names}{names of the groups.}

\item{excl}{vector indicating excluded modalities.}
}
\value{
list containing group results.
}
\description{
Calculate group contributions.
}
