% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gda-optimise-df.R
\name{gda_optimise_df}
\alias{gda_optimise_df}
\title{Optimise data frame for Geometric Data Analysis}
\usage{
gda_optimise_df(df_name, mod_excl = NA, prop_na_excl = 0.2,
  rename_na = FALSE)
}
\arguments{
\item{df_name}{name of the data frame to optimise.}

\item{mod_excl}{specify, which modalities should excluded.}

\item{prop_na_excl}{overall level to exclude specified modalities.}

\item{rename_na}{rename NA with label}
}
\value{
optimised data frame.
}
\description{
Optimise data frame for Geometric Data Analysis
}
