% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-gda-trajectory.R
\name{get_gda_trajectory}
\alias{get_gda_trajectory}
\title{Extract trajectory data.}
\usage{
get_gda_trajectory(res_gda, time_point_names = NULL)
}
\arguments{
\item{res_gda}{GDA result.}

\item{time_point_names}{name of the separated time points.}
}
\value{
list containing time point separated ind coord and time point names.
}
\description{
Extract trajectory data.
}

