% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-index-mod.R
\name{get_index_mod}
\alias{get_index_mod}
\title{Exctract index of specific modalities.}
\usage{
get_index_mod(df_gda, pattern = "Fehlender Wert")
}
\arguments{
\item{df_gda}{GDA optimised data frame.}

\item{pattern}{search pattern (regular expression).}
}
\value{
indices of modalities.
}
\description{
Exctract index of specific modalities.
}
