% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-mca-var-corr.R
\name{get_mca_var_corr}
\alias{get_mca_var_corr}
\title{Reshapes MCA results.}
\usage{
get_mca_var_corr(res_mca, axes = 1:2)
}
\arguments{
\item{res_mca}{MCA result.}

\item{axes}{axes selection.}
}
\value{
data frame with MCA variable names, types, coords and eta2 values.
}
\description{
Reshapes MCA results.
}
