% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-coord.R
\name{get_path_coord}
\alias{get_path_coord}
\title{Extract coords of categories to concat.}
\usage{
get_path_coord(res_gda_quali, var, var_levels = NULL, exclude = NULL)
}
\arguments{
\item{res_gda_quali}{GDA result.}

\item{var}{variable names.}

\item{var_levels}{variable categories to concat.}

\item{exclude}{categories to exclude.}
}
\description{
Extract coords of categories to concat.
}

