% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-places-chronology-time-pattern.R
\name{get_places_chronology_time_pattern}
\alias{get_places_chronology_time_pattern}
\title{Reshpape place chronology time pattern data.}
\usage{
get_places_chronology_time_pattern(data, id = "all", weekday = "all")
}
\arguments{
\item{data}{data frame, which contains place chronology time pattern data.}

\item{id}{vector, which contains questionnaire ids. Use "all" if you want to plot all ids.}

\item{weekday}{vector, which contains days to plot.}
}
\value{
reshaped data frame for further visualization.
}
\description{
Reshpape place chronology time pattern data.
}
