% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-time-pattern.R
\name{get_time_pattern}
\alias{get_time_pattern}
\title{Reshape time pattern data.}
\usage{
get_time_pattern(data, id = "all", reshape_data = TRUE)
}
\arguments{
\item{data}{}
}
\value{
reshaped data frame for further visualization.
}
\description{
Reshape time pattern data.
}

