% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-time-pattern-profile.R
\name{get_time_pattern_profile}
\alias{get_time_pattern_profile}
\title{Reshape time pattern profile data frame.}
\usage{
get_time_pattern_profile(data_tp, id = "all")
}
\arguments{
\item{data_tp}{data frame containing questionnaire_id, kml3d results and time pattern data.}

\item{id}{time pattern number.}
}
\value{
Reshaped data frame.
}
\description{
Reshape time pattern profile data frame.
}
