% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-time-pattern-series.R
\name{get_time_pattern_series}
\alias{get_time_pattern_series}
\title{Reshape time pattern series data.}
\usage{
get_time_pattern_series(data_tp)
}
\arguments{
\item{data_tp}{data frame including questionnaire_id, kml3d results and time pattern data.}
}
\value{
Reshaped data frame.
}
\description{
Reshape time pattern series data.
}
