% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified-rates.R
\name{modified_rates}
\alias{modified_rates}
\title{Calculate modified rates}
\usage{
modified_rates(mca_res)
}
\arguments{
\item{mca_res}{FactoMineR MCA object.}
}
\value{
Modified rates as tibble.
}
\description{
Calculate modified rates
}
