% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sMCA.R
\name{plot.sMCA}
\alias{plot.sMCA}
\title{Plot specific Multiple Correspondance Analysis results.}
\usage{
\method{plot}{sMCA}(x, axes = c(1, 2), choix = c("ind", "var",
  "quanti.sup"), xlim = NULL, ylim = NULL, invisible = c("none", "ind",
  "var", "ind.sup", "quali.sup", "quanti.sup"), col.ind = "blue",
  col.var = "red", col.quali.sup = "darkgreen", col.ind.sup = "darkblue",
  col.quanti.sup = "blue", label = c("all", "none", "ind", "var", "ind.sup",
  "quali.sup", "quanti.sup"), title = NULL, habillage = "none",
  palette = NULL, autoLab = c("auto", "yes", "no"), new.plot = FALSE,
  select = NULL, selectMod = NULL, unselect = 0.7, shadowtext = FALSE,
  ...)
}
\arguments{
\item{...}{}
}
\value{
Retutns a standard plot of sMCA results.
}
\description{
Plot specific Multiple Correspondance Analysis results.
}

