% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sMFA.R
\name{plot.sMFA}
\alias{plot.sMFA}
\title{Plot specific Multiple Factor Analysis results.}
\usage{
\method{plot}{sMFA}(x, axes = c(1, 2), choix = c("ind", "var", "group",
  "axes", "freq"), ellipse = NULL, ellipse.par = NULL, lab.grpe = TRUE,
  lab.var = TRUE, lab.ind = TRUE, lab.par = FALSE, lab.col = TRUE,
  habillage = "group", col.hab = NULL, invisible = c("none", "ind",
  "ind.sup", "quanti", "quanti.sup", "quali", "quali.sup", "row", "row.sup",
  "col", "col.sup"), partial = NULL, lim.cos2.var = 0, chrono = FALSE,
  xlim = NULL, ylim = NULL, title = NULL, palette = NULL,
  autoLab = c("auto", "yes", "no"), new.plot = FALSE, select = NULL,
  unselect = 0.7, shadowtext = FALSE, ...)
}
\arguments{
\item{...}{}
}
\value{
Returns the standard plots of sMFA results.
}
\description{
Plot specific Multiple Factor Analysis results.
}

