% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.spHMFA.R
\name{plot.spHMFA}
\alias{plot.spHMFA}
\title{Plot Secific Hierarchical Multiple Factor Analysis.}
\usage{
\method{plot}{spHMFA}(x, axes = c(1, 2), num = 6, choix = "ind",
  plot.nodes = FALSE, select = NULL, unselect = 0.7, lab.grpe = TRUE,
  lab.var = TRUE, lab.ind.moy = TRUE, invisible = NULL,
  lim.cos2.var = 0, xlim = NULL, ylim = NULL, cex = 1, title = NULL,
  new.plot = FALSE, ...)
}
\arguments{
\item{...}{}
}
\value{
Returns standard plots of spHMFA results.
}
\description{
Plot Secific Hierarchical Multiple Factor Analysis.
}

