% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-barplot.R
\name{plot_barplot}
\alias{plot_barplot}
\title{Visualize a barplot.}
\usage{
plot_barplot(dfname, xlab = "", ylab = "", title = "", sort = FALSE,
  labels_inline = FALSE, amount = FALSE, rotate_x_axis_text = FALSE,
  textsize = 20, titlesize = 25, labelsize = 8, include_na = TRUE,
  digits = 0, ylim = NA, xlim = NA, myriad = TRUE)
}
\arguments{
\item{dfname}{categorical variable.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{title}{plot title.}

\item{sort}{sort bars (boolean).}

\item{labels_inline}{inline labels (boolean).}

\item{amount}{include total amount of observations (boolean).}

\item{rotate_x_axis_text}{rotate bars (boolean).}

\item{textsize}{sife of axes texts.}

\item{titlesize}{title text size.}

\item{labelsize}{label text size.}

\item{include_na}{show NAs or not (boolean).}

\item{digits}{amount of label value digits.}

\item{ylim}{y-axis range.}

\item{xlim}{y-axis range.}

\item{myriad}{use Myriad Pro font.}
}
\value{
ggplot2 barplot.
}
\description{
Visualize a barplot.
}
