% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-places-chronology.R
\name{plot_places_chronology}
\alias{plot_places_chronology}
\title{Plot single or multiple places chronologies in different scales.}
\usage{
plot_places_chronology(data, id = "all", map, weekday = "all",
  timepattern = TRUE, distinct_places = FALSE,
  print_prop_duration = FALSE, legends = FALSE, structure = FALSE,
  title = "Orte Chronologie")
}
\arguments{
\item{structure}{}
}
\description{
Plot single or multiple places chronologies in different scales.
}

