% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-places-chronology-time-pattern.R
\name{plot_places_chronology_time_pattern}
\alias{plot_places_chronology_time_pattern}
\title{Plot place chronology time pattern data.}
\usage{
plot_places_chronology_time_pattern(data, id = "all", weekday = "all",
  graph = TRUE, print_prop_duration = TRUE, legend = TRUE,
  bar_width = 1, ncol = 3)
}
\arguments{
\item{data}{data frame, which contains place chronology time pattern data.}

\item{id}{vector, which contains questionnaire ids. Use "all" if you want to plot all ids.}

\item{weekday}{weekday vector, which contains days to plot.}

\item{graph}{whether to plot or not to plot the praph (boolean)}

\item{print_prop_duration}{whether to print or not to print prop duration data.}

\item{legend}{show or hide legends (boolean).}

\item{bar_width}{specify the width of the bars.}

\item{ncol}{number of cols, if there are multiple plots (facets).}
}
\value{
ggplot2 visualization of place chronology time pattern data.
}
\description{
Plot place chronology time pattern data.
}

