% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-time-pattern.R
\name{plot_time_pattern}
\alias{plot_time_pattern}
\title{Plot ingle or multiple time pattern.}
\usage{
plot_time_pattern(data, id = "all", ncol = 3, reshape_data = TRUE)
}
\arguments{
\item{ncol}{}
}
\description{
Plot ingle or multiple time pattern.
}

