% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sMCA.R
\name{print.sMCA}
\alias{print.sMCA}
\title{Print results of sMCA.}
\usage{
\method{print}{sMCA}(x, file = NULL, sep = ";", ...)
}
\arguments{
\item{...}{}
}
\value{
Returns a list of sMCA results.
}
\description{
Print results of sMCA.
}

