% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sMFA.R
\name{print.sMFA}
\alias{print.sMFA}
\title{Print summaory of sMFA results.}
\usage{
\method{print}{sMFA}(x, file = NULL, sep = ";", ...)
}
\arguments{
\item{...}{}
}
\value{
List of sMFA results.
}
\description{
Print summaory of sMFA results.
}

