% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sHMFA.R
\name{print.spHMFA}
\alias{print.spHMFA}
\title{Print sHMFA summary.}
\usage{
\method{print}{spHMFA}(x, file = NULL, sep = ";", ...)
}
\arguments{
\item{...}{}
}
\value{
List of sHMFA results.
}
\description{
Print sHMFA summary.
}

