% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.spMCA.R
\name{print.spMCA}
\alias{print.spMCA}
\title{Print results of spMCA.}
\usage{
\method{print}{spMCA}(x, file = NULL, sep = ";", ...)
}
\arguments{
\item{...}{}
}
\value{
Returns a list of spMCA results.
}
\description{
Print results of spMCA.
}

