% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.spMFA.R
\name{print.spMFA}
\alias{print.spMFA}
\title{Print summaory of spMFA results.}
\usage{
\method{print}{spMFA}(x, file = NULL, sep = ";", ...)
}
\arguments{
\item{...}{}
}
\value{
List of spMFA results.
}
\description{
Print summaory of spMFA results.
}

