% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCA.R
\name{sCA}
\alias{sCA}
\title{Perform specific Correspondance Analysis}
\usage{
sCA(X, ncp = 5, row.sup = NULL, col.sup = NULL, quanti.sup = NULL,
  quali.sup = NULL, graph = TRUE, axes = c(1, 2), row.w = NULL,
  excl = NULL)
}
\arguments{
\item{excl}{}
}
\value{
Returns a list with sCA results.
}
\description{
Perform specific Correspondance Analysis
}

