% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sHMFA.R
\name{sHMFA}
\alias{sHMFA}
\title{Perform specific Hierarchical Multiple Factor Analysis}
\usage{
sHMFA(X, H, type = rep("s", length(H[[1]])), excl = NULL, ncp = 5,
  graph = TRUE, axes = c(1, 2), name.group = NULL)
}
\arguments{
\item{name.group}{}
}
\value{
Returns a list of sHMFA results.
}
\description{
Perform specific Hierarchical Multiple Factor Analysis
}

