% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sMFA.R
\name{sMFA}
\alias{sMFA}
\title{Performs a specific Multiple Factor Analysis.}
\usage{
sMFA(base, group, type = rep("s", length(group)), excl = NULL,
  ind.sup = NULL, ncp = 5, name.group = NULL, num.group.sup = NULL,
  graph = TRUE, weight.col.mfa = NULL, row.w = NULL, axes = c(1, 2),
  tab.comp = NULL)
}
\arguments{
\item{tab.comp}{}
}
\value{
Returns a list of sMFA results.
}
\description{
Performs a specific Multiple Factor Analysis.
}

