% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCA.R
\name{spCA}
\alias{spCA}
\title{Perform "Specific Correspondance Analysis"}
\usage{
spCA(X, ncp = 5, row.sup = NULL, col.sup = NULL, quanti.sup = NULL,
  quali.sup = NULL, graph = TRUE, axes = c(1, 2), row.w = NULL,
  excl = NULL)
}
\arguments{
\item{excl}{}
}
\value{
Returns a list with CA results.
}
\description{
Perform "Specific Correspondance Analysis"
}

