% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific-catdes.R
\name{specific_catdes}
\alias{specific_catdes}
\title{Category description}
\usage{
specific_catdes(donnee, num.var, proba = 0.05, row.w = NULL)
}
\arguments{
\item{donnee}{}

\item{num.var}{}

\item{proba}{}

\item{row.w}{}
}
\value{
Multivariate statistic descriptions.
}
\description{
Category description
}

