% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific-condes.R
\name{specific_condes}
\alias{specific_condes}
\title{Perform ANOVA and correlation calculations}
\usage{
specific_condes(donnee, num.var, weights = NULL, proba = 0.05)
}
\arguments{
\item{proba}{}
}
\value{
Multivariate statistis for specific_dimdesc
}
\description{
Perform ANOVA and correlation calculations
}

