% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific-dimdesc.R
\name{specific_dimdesc}
\alias{specific_dimdesc}
\title{Dimension description of specific component analysis.}
\usage{
specific_dimdesc(res, axes = 1:3, proba = 0.05)
}
\arguments{
\item{proba}{}
}
\value{
list of representative kategories and modalities.
}
\description{
Dimension description of specific component analysis.
}

