% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spHCPC.R
\name{spHCPC}
\alias{spHCPC}
\title{Performs a Specific Hierarchical Cluster of Principal Components.}
\usage{
spHCPC(res, nb.clust = 0, consol = TRUE, iter.max = 10, min = 3,
  max = NULL, metric = "euclidean", method = "ward", order = TRUE,
  graph.scale = "inertia", nb.par = 5, graph = TRUE, proba = 0.05,
  cluster.CA = "rows", kk = Inf, ...)
}
\arguments{
\item{...}{}
}
\value{
Returns a list of spHCPC results.
}
\description{
Performs a Specific Hierarchical Cluster of Principal Components.
}

