% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spHMFA.R
\name{spHMFA}
\alias{spHMFA}
\title{Implementation of spHMFA. Use a list to exlcude junk modalities.}
\usage{
spHMFA(X, H, type = rep("s", length(H[[1]])), excl = NULL, ncp = 5,
  graph = TRUE, axes = c(1, 2), name.group = NULL)
}
\arguments{
\item{X}{a data.frame.}

\item{H}{a list with one vector for each hierarchical level; in each vector the number of variables or the number of group constituting the group.}

\item{type}{the type of variables in each group in the first partition; three possibilities: "c" or "s" for quantitative variables (the difference is that for "s", the variables are scaled in the program), "n" for categorical variables; by default, all the variables are quantitative and the variables are scaled unit.}

\item{excl}{an argument that may possible to exclude categories of active variables of categorical variable groups. NULL by default, it is a list with indexes of categories that are excluded per group.}

\item{ncp}{number of dimensions kept in the results (by default 5).}

\item{graph}{boolean, if TRUE a graph is displayed.}

\item{axes}{a length 2 vector specifying the components to plot.}

\item{name.group}{a list of vector containing the name of the groups for each level of the hierarchy (by default, NULL and the group are named L1.G1, L1.G2 and so on).}
}
\value{
HMFA like results.
}
\description{
Implementation of spHMFA. Use a list to exlcude junk modalities.
}

