% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spMCA.R
\name{spMCA}
\alias{spMCA}
\title{Performs a Specific Multiple Correspondance Analysis.}
\usage{
spMCA(X, ncp = 5, excl = NULL, ind.sup = NULL, quanti.sup = NULL,
  quali.sup = NULL, graph = TRUE, level.ventil = 0, axes = c(1, 2),
  row.w = NULL, method = "Indicator", na.method = "NA", tab.disj = NULL)
}
\arguments{
\item{tab.disj}{}
}
\value{
Returns a list of spMCA results.
}
\description{
Performs a Specific Multiple Correspondance Analysis.
}

