% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPCA.R
\name{spPCA}
\alias{spPCA}
\title{Performs a Specific Principal Component Analysis}
\usage{
spPCA(X, scale.unit = TRUE, ncp = 5, ind.sup = NULL, quanti.sup = NULL,
  structuring.factors = NULL, quali.sup = NULL, row.w = NULL,
  col.w = NULL, graph = TRUE, axes = c(1, 2))
}
\arguments{
\item{axes}{}
}
\value{
Returns a list of spPCA results.
}
\description{
Performs a Specific Principal Component Analysis
}

