% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supvar-crossing-stats.R
\name{supvar_crossing_stats}
\alias{supvar_crossing_stats}
\title{Calculate crossed variables double breakdown of variance.}
\usage{
supvar_crossing_stats(res_gda, var_quali_df, var_quali, impute = TRUE,
  axes = 1:2)
}
\arguments{
\item{res_gda}{GDA result.}

\item{var_quali_df}{the supplementary data frame.}

\item{var_quali}{crossed supplementary variable (vector separated by "_").}

\item{impute}{impute missing data (boolean).}

\item{axes}{the GDA dimensions to calculate.}
}
\value{
Returns a list:
\item{var}{double breakdown of variance.}
\item{reg}{result of linear regressions.}
}
\description{
Calculate crossed variables double breakdown of variance.
}

