% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supvar-stats.R
\name{supvar_stats}
\alias{supvar_stats}
\title{Calculate results for supplementary variables.}
\usage{
supvar_stats(res_gda, df_var_quali, var_quali_name)
}
\arguments{
\item{res_gda}{GDA result.}

\item{df_var_quali}{data frame of one quali variable.}

\item{var_quali_name}{name if quali variable.}
}
\value{
Returns a list:
\item{weight}{numeric vector of categories weights}
\item{cord}{data frame of categories coordinates}
\item{cos2}{data frame of categories square cosine}
\item{var}{data frame of categories within variances, variance between and within categories and variable square correlation ratio (eta2)}
\item{v.test}{data frame of categories test-values}
}
\description{
Calculate results for supplementary variables.
}

