% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-theme.R
\name{add_theme}
\alias{add_theme}
\title{Optimise ggplot2 plot.}
\usage{
add_theme(plot, font_family = "Myriad Pro")
}
\arguments{
\item{plot}{ggplot2 object.}

\item{font_family}{plot overall font family.}
}
\value{
ggplot2 geoms.
}
\description{
Optimise ggplot2 plot.
}
