% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-load.R
\name{db_load}
\alias{db_load}
\title{Save data frame as SQLite table.}
\usage{
db_load(remote_db_path, copy = TRUE)
}
\arguments{
\item{remote_db_path}{path to the remote database (SQLite).}

\item{copy}{copy bring data into an R (boolean).}
}
\value{
The SQLite database.
}
\description{
Save data frame as SQLite table.
}

