% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-save.R
\name{db_save}
\alias{db_save}
\title{Saves data into a SQLite database.}
\usage{
db_save(remote_db_path, source_df)
}
\arguments{
\item{remote_db_path}{path to the remote database (SQLite).}

\item{source_df}{source data frame (should be a dplyr tbl object).}
}
\value{
SQLite database file.
}
\description{
Saves data into a SQLite database.
}

