% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-conc-ellipse.R
\name{fviz_gda_conc_ellipse}
\alias{fviz_gda_conc_ellipse}
\title{Title}
\usage{
fviz_gda_conc_ellipse(res_gda, level = 0.8647, alpha = 0.1,
  colour = "black", linetype = "dashed", axes = 1:2)
}
\arguments{
\item{res_gda}{GDA result.}

\item{level}{ellipse level (default 86.47\%).}

\item{alpha}{opacity level (default 0.1).}

\item{colour}{ellipse fill colour.}

\item{linetype}{ellipse edge linetype.}

\item{axes}{the GDA dimensions to plot.}
}
\value{
ggplot2 GDA visualisation with concentration ellipse.
}
\description{
Title
}

