% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-interaction.R
\name{fviz_gda_interaction}
\alias{fviz_gda_interaction}
\title{Visualize interaction cloud.}
\usage{
fviz_gda_interaction(res_gda, df_var_quali, var_quali,
  title = "MCA quali interaction effects", mean_alpha = 0.75,
  path_linetype = "solid", path_size = 1, path_colour = "black",
  scale_mean_points = TRUE, axes = 1:2, palette = "Set1",
  path_alpha = 1, myriad = TRUE, impute = TRUE, variable = "both",
  plot_modif_rates = TRUE)
}
\arguments{
\item{res_gda}{MCA result.}

\item{df_var_quali}{crossed variable data frame.}

\item{var_quali}{name of crossed supplementary variable.}

\item{title}{plot title.}

\item{mean_alpha}{alpha of the mean point.}

\item{path_linetype}{linetype of concat path.}

\item{path_size}{size of concat path.}

\item{path_colour}{colour of concat path.}

\item{scale_mean_points}{scale mean points (boolean).}

\item{axes}{axes to plot.}

\item{palette}{used colour brewer palette.}

\item{path_alpha}{opacity of the path.}

\item{myriad}{use myriad font or not (boolean).}

\item{impute}{use imputation to handle missing data.}

\item{variable}{which diagram to plot (vector containing 1, 2 or "both").}

\item{plot_modif_rates}{plot modified rates instead of eigenvalue percentage (boolean).}
}
\value{
ggplot2 interaction cloud visualizsation.
}
\description{
Visualize interaction cloud.
}
