% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-quali-ellipses.R
\name{fviz_gda_quali_ellipses}
\alias{fviz_gda_quali_ellipses}
\title{Title}
\usage{
fviz_gda_quali_ellipses(res_gda, df_var_quali, var_quali_name,
  title = "MCA quali var ellipses", facet = TRUE, path_mean = FALSE,
  alpha.point = 0.75, path.linetype = "solid", scale_mean_points = TRUE,
  hcpc = FALSE, axes = 1:2, palette = "Set1")
}
\arguments{
\item{res_gda}{GDA result.}

\item{df_var_quali}{data frame of one quali variable.}

\item{var_quali_name}{name if quali variable.}

\item{title}{plot title.}

\item{facet}{whether facet ellipses or not (boolean).}

\item{path_mean}{plot mean point path (boolean). Only possible if facet is FALSE.}

\item{alpha.point}{opacity of mean points.}

\item{path.linetype}{linetype of mean point path.}

\item{scale_mean_points}{scale mean point size in respect of the group size (boolean).}

\item{hcpc}{add ellipses to HCPC results (boolean).}

\item{axes}{the GDA dimensions to plot.}

\item{palette}{Colour brewer scale.}
}
\value{
ggplo2 visualization with concentration and quali var ellipses.
}
\description{
Title
}

