% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-quali-ellipses.R
\name{fviz_gda_quali_ellipses}
\alias{fviz_gda_quali_ellipses}
\title{Title}
\usage{
fviz_gda_quali_ellipses(res_gda, df_var_quali, var_quali,
  title = "MCA quali var ellipses", facet = TRUE, alpha_point = 0.75,
  conc_linetype = "solid", conf_linetype = "solid",
  scale_mean_points = TRUE, axes = 1:2, palette = "Set1", myriad = TRUE,
  impute = TRUE, concentration_ellipses = TRUE,
  confidence_ellipses = TRUE, conf_colour = FALSE)
}
\arguments{
\item{res_gda}{GDA (MCA, MFA) result (rownames have to be individual questionnaire IDs).}

\item{df_var_quali}{data frame of one quali variable.}

\item{var_quali}{name if quali variable.}

\item{title}{plot title.}

\item{facet}{whether facet ellipses or not (boolean).}

\item{alpha_point}{opacity of individual points.}

\item{conc_linetype}{linetype of concentration ellipse.}

\item{conf_linetype}{linetype of confidence ellipse.}

\item{scale_mean_points}{scale mean point size in respect of the group size (boolean).}

\item{axes}{the GDA dimensions to plot.}

\item{palette}{Colour brewer scale.}

\item{myriad}{use Myriad Pro font family (boolean).}

\item{impute}{impute missing data (boolean).}

\item{concentration_ellipses}{plot concentration ellipse (boolean).}

\item{confidence_ellipses}{plot confidence ellipses (boolean).}

\item{conf_colour}{colour confidence ellipses (boolean).}
}
\value{
ggplo2 visualization with concentration and quali var ellipses.
}
\description{
Title
}

