% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-quali-supvar.R
\name{fviz_gda_quali_supvar}
\alias{fviz_gda_quali_supvar}
\title{Visualize supplementary variables.}
\usage{
fviz_gda_quali_supvar(res_gda, df_var_quali, var_quali,
  title = "MCA quali var structure", path = FALSE, linetype = "solid",
  axes = 1:2, scale_point = TRUE, size_point = 3, scale_text = FALSE,
  size_text = 3, palette = "Set1", myriad = TRUE, impute = TRUE,
  plot_modif_rates = TRUE, impute_ncp = 2, relevel = NULL,
  plot_eta2 = TRUE, axis_lab_name = "Achse")
}
\arguments{
\item{res_gda}{GDA result.}

\item{df_var_quali}{crossed variable data frame.}

\item{var_quali}{crossed variable name.}

\item{title}{plot title.}

\item{path}{plot path (boolean).}

\item{linetype}{specify path linetype.}

\item{axes}{which axes should be plotted.}

\item{scale_point}{scale points by weight (boolean).}

\item{size_point}{define point size.}

\item{scale_text}{scale text by weight (boolean).}

\item{size_text}{define text size.}

\item{palette}{RColorBrewer palette.}

\item{myriad}{use Myriad Pro font family (boolean).}

\item{impute}{impute missing data (boolean).}

\item{plot_modif_rates}{plot modified rates instead of eigenvalue percentage (boolean).}

\item{impute_ncp}{number of dimensions to predict missing values.}

\item{relevel}{character vector containing new level order.}

\item{plot_eta2}{plot eta2 value per axis (boolean).}

\item{axis_lab_name}{name of axis label.}
}
\value{
ggplot2 visualization of supplementary variables.
}
\description{
Visualize supplementary variables.
}
