% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-structure.R
\name{fviz_gda_structure}
\alias{fviz_gda_structure}
\title{Concat crossed variables.}
\usage{
fviz_gda_structure(res_gda, df_var_quali, var_quali_name,
  title = "MCA quali structure effects", mean.alpha = 0.75,
  path.linetype = "solid", path.size = 1, path.colour = NULL,
  scale_mean_points = TRUE, axes = 1:2, palette = "Set1",
  mod_names = NULL, mod_level_order = NULL, path.alpha = 0.7,
  myriad = TRUE)
}
\arguments{
\item{res_gda}{GDA result.}

\item{df_var_quali}{crossed variable data frame.}

\item{var_quali_name}{crossed variable name.}

\item{title}{plot title.}

\item{mean.alpha}{alpha of the mean point.}

\item{path.linetype}{linetype of concat path.}

\item{path.size}{size of concat path.}

\item{path.colour}{colour of concat path.}

\item{scale_mean_points}{scale mean points (boolean).}

\item{axes}{axes to plot.}

\item{palette}{used colour brewer palette.}

\item{mod_names}{names of the crossed categories.}

\item{mod_level_order}{order of the crossed modalities.}
}
\value{
ggplot2 visualizsation.
}
\description{
Concat crossed variables.
}

