% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-structure.R
\name{fviz_gda_structure}
\alias{fviz_gda_structure}
\title{Visualize additive cloud.}
\usage{
fviz_gda_structure(res_gda, df_var_quali, var_quali,
  title = "MCA quali structure effects", scale_mean_points = TRUE,
  axes = 1:2, palette = "Set1", impute = TRUE, myriad = TRUE,
  cloud = "both", plot_modif_rates = TRUE, axis_lab_name = "Achse")
}
\arguments{
\item{res_gda}{MCA result.}

\item{df_var_quali}{crossed variable data frame.}

\item{var_quali}{name of quali variable.}

\item{title}{plot title.}

\item{scale_mean_points}{scale points (boolean).}

\item{axes}{which axis to plot.}

\item{palette}{colour palette (boolean).}

\item{impute}{impute missing data (boolean).}

\item{myriad}{use Myriad Pro font (boolean).}

\item{cloud}{which cloud should be plotted (string: real, fitted, both, deviation)}

\item{plot_modif_rates}{plot modified rates instead of eigenvalue percentage (boolean).}

\item{axis_lab_name}{name of axis label.}
}
\value{
ggplot2 visualization of additive cloud.
}
\description{
Visualize additive cloud.
}
