% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-trajectory-conc.R
\name{fviz_gda_trajectory_conc}
\alias{fviz_gda_trajectory_conc}
\title{Title}
\usage{
fviz_gda_trajectory_conc(res_gda, time_point_names = NULL, myriad = TRUE,
  axes = 1:2)
}
\description{
Title
}

