% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-trajectory-ellipses.R
\name{fviz_gda_trajectory_ellipses}
\alias{fviz_gda_trajectory_ellipses}
\title{Visualization of trajectory structuring factor ellipses.}
\usage{
fviz_gda_trajectory_ellipses(res_gda, df_var_quali, var_quali, axes = 1:2,
  myriad = TRUE, impute = TRUE, time_point_names = NULL,
  ind_points = TRUE, concentration_ellipse = TRUE,
  title = "Trajectory individuals structuring factors ellipse plot",
  plot_modif_rates = TRUE, alpha = 0.15, select = NULL,
  axis_lab_name = "Achse")
}
\arguments{
\item{res_gda}{MCA result (rownames have to be questionnaire IDs including time number, e.g. 87654_1).}

\item{df_var_quali}{data frame containing one qualitative variable (with IDs as rownames).}

\item{var_quali}{name of the structuring variable.}

\item{axes}{the axes to plot.}

\item{myriad}{use Myriad Pro font (boolean).}

\item{impute}{use imputation for missing data.}

\item{time_point_names}{vector containing the name of the time points.}

\item{ind_points}{show individuals (boolean).}

\item{concentration_ellipse}{plot confidence ellipses (boolean).}

\item{title}{title of the plot.}

\item{plot_modif_rates}{plot modified rates instead of eigenvalue percentage (boolean).}

\item{alpha}{ellipse fill alpha.}

\item{select}{choose time point.}

\item{axis_lab_name}{name of axis label.}
}
\value{
ggplot2 visualization.
}
\description{
Visualization of trajectory structuring factor ellipses.
}
