% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-var-axis.R
\name{fviz_gda_var_axis}
\alias{fviz_gda_var_axis}
\title{Visualize specific contributing modalities.}
\usage{
fviz_gda_var_axis(res_gda, axis = 1, contrib = "auto",
  title = "GDA axis high contribution modalities", axes = 1:2,
  group = NULL, group_names = NULL, group_style = "both", textsize = 4,
  colour_palette = "Dark2", myriad = TRUE, plot_modif_rates = TRUE)
}
\arguments{
\item{res_gda}{GDA result data frame.}

\item{axis}{dimension to be filtered.}

\item{contrib}{"auto" calculates the optimal modalities to show (based on the basic criterion). Otherwise define an amount of modalities to plot.}

\item{title}{plot title.}

\item{axes}{the GDA dimensions to plot.}

\item{group}{vector containing group definition.}

\item{group_names}{names of the groups.}

\item{group_style}{style to plot (vector containing "shape", "colour" or "both).}

\item{textsize}{size of the text.}

\item{colour_palette}{name of the used colour palette.}

\item{myriad}{use Myriad Pro font (boolean).}

\item{plot_modif_rates}{plot modified rates instead of eigenvalue percentage (boolean).}
}
\value{
ggplot2 visualization containing selected modalities.
}
\description{
Visualize specific contributing modalities.
}

