% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gda-describe-axis.R
\name{gda_describe_axis}
\alias{gda_describe_axis}
\title{Calculate axis contributions.}
\usage{
gda_describe_axis(res_gda, axis = 1, contrib = "auto")
}
\arguments{
\item{res_gda}{MCA result.}

\item{axis}{which axis to calculate.}

\item{contrib}{"auto" calculates the optimal modalities to show (based on the basic criterion). Otherwise define an amount of modalities to plot.}
}
\value{
list containing axis contribution results.
}
\description{
Calculate axis contributions.
}

