% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-places-chronology.R
\name{plot_places_chronology}
\alias{plot_places_chronology}
\title{Plot single or multiple places chronologies in different scales.}
\usage{
plot_places_chronology(data, id = "all", map, weekday = "all",
  timepattern = TRUE, distinct_places = FALSE, size_range = c(3, 15),
  shape_path = 1, colour_path = "black", size_path = 2,
  alpha_path = 0.75, linetype_path = "solid", force_repel = 1,
  print_prop_duration = FALSE, legends = FALSE, structure = FALSE,
  title = "Orte Chronologie")
}
\arguments{
\item{title}{}
}
\description{
Plot single or multiple places chronologies in different scales.
}

