% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-places-chronology.R
\name{plot_places_chronology}
\alias{plot_places_chronology}
\title{Plot single or multiple place chronologies in different scales.}
\usage{
plot_places_chronology(data, id = "all", weekday = "all", map = NULL,
  size_range = c(3, 15), shape_path = 1, colour_path = "black",
  size_path = 2, alpha_path = 0.75, linetype_path = "solid",
  force_repel = 1, legend = FALSE, structure = FALSE,
  map_extent = "panel", title = "Orte Chronologie", axis_label = FALSE,
  xlim = NULL, ylim = NULL, graph = TRUE, ncol = 3,
  unique_places = FALSE, print_place_duration = FALSE,
  activity_duration_overall = TRUE, facet_scales = "fixed",
  point_padding = unit(1e-06, "lines"), myriad = TRUE,
  exclude_sleep = TRUE)
}
\arguments{
\item{data}{a data frame (columns: ID, day, duration, place, address, lon, lat, prop_duration).}

\item{id}{vector, which contains questionnaire ids. Choosa "all" to compute all ids.}

\item{weekday}{vactor, which contains the weekday to plot.}

\item{map}{a ggmap map object.}

\item{size_range}{specify the size for visualizatipn of duration.}

\item{shape_path}{specify the shape of the path line. Higher values result in smoother paths.}

\item{colour_path}{sepcify the path line colour.}

\item{size_path}{specify the path line size.}

\item{alpha_path}{specify the path line alpha value [0:1].}

\item{linetype_path}{specify the linetype of the path line.}

\item{force_repel}{specify how heavy the repel algorithmn should be.}

\item{legend}{show or hide legends (boolean).}

\item{structure}{show or hide map background (boolean).}

\item{map_extent}{how much of the plot should the map take up? "normal", "device", or "panel" (default).}

\item{title}{title of the plot.}

\item{axis_label}{show or hide axis labels (boolean).}

\item{xlim}{specify plot x limits.}

\item{ylim}{spcify plot y limits.}

\item{graph}{whether to plot or not to plot the praph (boolean).}

\item{ncol}{number of cols, if there are multiple plots (facets).}

\item{unique_places}{plot only unique places. Useful to visualize place related overall structure.}

\item{print_place_duration}{print place overall duration (hours).}

\item{activity_duration_overall}{wheter plot activities per week or for all weeks at once (boolean).}

\item{facet_scales}{should Scales be fixed ("fixed", the default), free ("free"), or free in one dimension ("free_x", "free_y").}

\item{point_padding}{Amount of padding around labeled point. Defaults to unit(0, "lines").}

\item{myriad}{use Myriad Pro font (boolean).}

\item{exclude_sleep}{exclude sleep duration (boolean).}
}
\value{
ggplot2 visualization of place chronology data.
}
\description{
Plot single or multiple place chronologies in different scales.
}

