% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-time-pattern.R
\name{plot_time_pattern}
\alias{plot_time_pattern}
\title{Plot ingle or multiple time pattern.}
\usage{
plot_time_pattern(data, ncol = 3)
}
\arguments{
\item{ncol}{}
}
\description{
Plot ingle or multiple time pattern.
}

