% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-time-pattern.R
\name{plot_time_pattern}
\alias{plot_time_pattern}
\title{Plot single or multiple time pattern.}
\usage{
plot_time_pattern(data, id = "all", ncol = 3, reshape_data = TRUE,
  print_prop_duration = TRUE)
}
\arguments{
\item{data}{data frame which contains time pattern data.}

\item{id}{vector which contains questionnaire ids.}

\item{ncol}{number of cols, if there are multiple plots (facets).}

\item{reshape_data}{whether reshape data or not.}

\item{print_prop_duration}{whether to print or not to print prop duration data.}
}
\value{
ggplot2 visualization of time pattern data.
}
\description{
Plot single or multiple time pattern.
}

