% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.spHMFA.R
\name{print.spHMFA}
\alias{print.spHMFA}
\title{Print spHMFA summary.}
\usage{
\method{print}{spHMFA}(x, file = NULL, sep = ";", ...)
}
\arguments{
\item{...}{}
}
\value{
List of spHMFA results.
}
\description{
Print spHMFA summary.
}

