% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sPCA.R
\name{sPCA}
\alias{sPCA}
\title{Performs a specific Principal Component Analysis}
\usage{
sPCA(X, scale.unit = TRUE, ncp = 5, ind.sup = NULL, quanti.sup = NULL,
  structuring.factors = NULL, quali.sup = NULL, row.w = NULL,
  col.w = NULL, graph = TRUE, axes = c(1, 2))
}
\arguments{
\item{axes}{}
}
\value{
Returns a list of sPCA results.
}
\description{
Performs a specific Principal Component Analysis
}

