% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spHMFA.R
\name{spHMFA}
\alias{spHMFA}
\title{Perform Specific Hierarchical Multiple Factor Analysis}
\usage{
spHMFA(X, H, type = rep("s", length(H[[1]])), excl = NULL, ncp = 5,
  graph = TRUE, axes = c(1, 2), name.group = NULL)
}
\arguments{
\item{name.group}{}
}
\value{
Returns a list of spHMFA results.
}
\description{
Perform Specific Hierarchical Multiple Factor Analysis
}

