% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supvar-crossing-stats.R
\name{supvar_crossing_stats}
\alias{supvar_crossing_stats}
\title{Calculate crossed variables double breakdown of variance.}
\usage{
supvar_crossing_stats(res_gda, quali_var, impute = TRUE)
}
\arguments{
\item{res_gda}{GDA result.}

\item{quali_var}{crossed variable (vector separated by "_").}

\item{impute}{handle missing values by imputation.}
}
\value{
Returns a list:
\item{between_var1}{between variance var1}
\item{var1_within_var2}{var1 within var2 variance}
\item{between_var2}{data frame of categories square cosine}
\item{var2_within_var1}{data frame of categories within variances, variance between and within categories and variable square correlation ratio (eta2)}
\item{additive}{data frame of categories test-values}
\item{interaction}{data frame of categories test-values}
\item{between_var1_var2}{data frame of categories test-values}
}
\description{
Calculate crossed variables double breakdown of variance.
}

