% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supvar-stats.R
\name{supvar_stats}
\alias{supvar_stats}
\title{Calculate results for supplementary variables.}
\usage{
supvar_stats(res_gda, var_quali_df, var_quali, impute = TRUE,
  impute_ncp = 2)
}
\arguments{
\item{res_gda}{GDA result.}

\item{var_quali_df}{the supplementary data frame.}

\item{var_quali}{supplementary variable name (string).}

\item{impute}{impute missing data (boolean).}

\item{impute_ncp}{number of dimensions to predict missing values.}
}
\value{
Returns a list:
\item{weight}{numeric vector of categories weights}
\item{cord}{data frame of categories coordinates}
\item{cos2}{data frame of categories square cosine}
\item{var}{data frame of categories within variances, variance between and within categories and variable square correlation ratio (eta2)}
\item{v.test}{data frame of categories test-values}
\item{supvar}{vector of the supplementary variable categories}
}
\description{
Calculate results for supplementary variables.
}
